Rename icolbutler-29-1 to icolbutler-29
Changed the name of the document for consistency: my idea was that the *branches* would be sub-labelled (1, 2, 3, etc.) but we'd want a single "master" version of each file on the master branch, with the simplest name possible.

Also: just testing to see what happens if I commit an edit directly to a branch someone else is working on! Will that break everything, or can he handle it "in the browser"? Let's find out!